Some changes.

+++ src/main/java/com/github/mcnaughtondesktop/CmaxCalculator.java
@@ -9,7 +9,7 @@ public class CmaxCalculator {
     private static final String TEMPLATE = "Cmax = max{max{%s}, %s / %s}} = max{%s, %s} = %s";
     private final List<Task> tasks;
     private final int machineCount;
-    private int result;
+    private double result;
     private String resultString;
 
     public CmaxCalculator(int machineCount, List<Task> tasks) {
@@ -19,17 +19,17 @@ public class CmaxCalculator {
     }
 
     private void calculate() {
-        int taskExecutionTimesSum = tasks.stream().mapToInt(Task::getDuration).sum();
-        int maxTaskDuration = tasks.stream().mapToInt(Task::getDuration).reduce(Integer::max).orElseThrow();
-        int averageDurationPerMachine = tasks.stream().mapToInt(Task::getDuration).sum() / machineCount;
-        result = Integer.max(maxTaskDuration, averageDurationPerMachine);
+        double taskExecutionTimesSum = tasks.stream().mapToDouble(Task::getDuration).sum();
+        double maxTaskDuration = tasks.stream().mapToDouble(Task::getDuration).reduce(Double::max).orElseThrow();
+        double averageDurationPerMachine = tasks.stream().mapToDouble(Task::getDuration).sum() / machineCount;
+        result = Double.max(maxTaskDuration, averageDurationPerMachine);
 
         String taskExecutionTimesString = tasks.stream().map(Task::getDuration).collect(Collectors.toList()).toString();
         String taskExecutionTimesStringTrimmed = taskExecutionTimesString.substring(1, taskExecutionTimesString.length() - 1);
         resultString = String.format(TEMPLATE, taskExecutionTimesStringTrimmed, taskExecutionTimesSum, machineCount, maxTaskDuration, averageDurationPerMachine, result);
     }
 
-    public int getResult() {
+    public double getResult() {
         return result;
     }
