+++ src/main/java/com/github/mcnaughtondesktop/model/McNaughtonResult.java
@@ -3,15 +3,15 @@ package com.github.mcnaughtondesktop.model;
 import java.util.List;
 
 public class McNaughtonResult {
-    private final Integer cmax;
+    private final Double cmax;
     private final List<Machine> machines;
 
-    public McNaughtonResult(Integer cmax, List<Machine> machines) {
+    public McNaughtonResult(Double cmax, List<Machine> machines) {
         this.cmax = cmax;
         this.machines = machines;
     }
 
-    public Integer getCmax() {
+    public Double getCmax() {
         return cmax;
     }
